+++ src/app/convenience-file-types.test.js
@@ -1,5 +1,6 @@
 import expect from 'expect';
 import {
+  expandExpressionMatrixZarr,
   expandRasterJson,
   expandRasterOmeZarr,
   expandGenesJson,
@@ -12,6 +13,23 @@ import {
 } from './convenience-file-types';
 
 describe('src/app/convenience-file-types.js', () => {
+  describe('expandExpressionMatrixZarr', () => {
+    it('expands expression-matrix.zarr', () => {
+      expect(expandExpressionMatrixZarr({
+        fileType: 'expression-matrix.zarr',
+        url: 'http://localhost:8000/expression-matrix.zarr',
+      })).toEqual([
+        {
+          fileType: 'obsFeatureMatrix.expression-matrix.zarr',
+          url: 'http://localhost:8000/expression-matrix.zarr',
+          coordinationValues: {
+            obsType: 'cell',
+            featureType: 'gene',
+          },
+        },
+      ]);
+    });
+  });
   describe('expandRasterJson', () => {
     it('expands raster.json', () => {
       expect(expandRasterJson({