+++ src/loaders/index.js
@@ -1,5 +1,5 @@
 export { default as JsonLoader } from './JsonLoader';
-export { default as MatrixZarrLoader } from './MatrixZarrLoader';
+export { default as MatrixZarrLoader } from './matrix-loaders/MatrixZarrLoader';
 export { default as ClustersJsonAsMatrixZarrLoader } from './clusters-json-loaders/ClustersJsonAsMatrixZarrLoader';
 export { default as GenesJsonAsMatrixZarrLoader } from './genes-json-loaders/GenesJsonAsMatrixZarrLoader';
 export { default as AnnDataLoaders } from './anndata-loaders';